coslidate logic
