+++ server/utils/limiter/extension.py
@@ -25,12 +25,11 @@ from limits.aio.strategies import FixedWindowRateLimiter, RateLimiter
 from limits.errors import StorageError
 from pydantic import BaseModel
 from utils.limiter.storage import ValkeyStorage
-from utils.responses.orjson import ORJSONResponse
+from utils.responses import ORJSONResponse
 
 if TYPE_CHECKING:
-    from core import Kanae
+    from core import Kanae, KanaeConfig
     from starlette.datastructures import MutableHeaders
-    from utils.config import KanaeConfig
     from utils.glide import GlideManager
 
 # Define an alias for the most commonly used type
