--- server/utils/responses/__init__.py
@@ -1,1 +1,0 @@
-HTTP_404_DETAIL = "Resource not found"
